+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+  - background_color as an option for displaying text
+  -
+### Changed
+  - modified the way to pass in text color
+    * takes an atom now instead of an `IO.ANSI`color function
 
 ## [0.6.2] - 2021-10-25
 ### Docuentation