+++ CHANGELOG.md
@@ -7,10 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
   - background_color as an option for displaying text
-  -
+
 ### Changed
   - modified the way to pass in text color
     * takes an atom now instead of an `IO.ANSI`color function
+  - refactored `select/2` into it's own module
 
 ## [0.6.2] - 2021-10-25
 ### Docuentation