+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [0.6.3] - 2021-11-28
 ### Added
   - use NimbleOptions to validate options
   - background_color as an option for displaying text
@@ -39,7 +42,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
   - Instead of taking a list of tuples for commands (`[{"command", Module"}]`), now take a Keyword list (`[command: Module]`)
 
-[Unreleased]: https://github.com/silbermm/prompt/compare/v0.6.2...HEAD
+[Unreleased]: https://github.com/silbermm/prompt/compare/v0.6.3...HEAD
+[0.6.3]: https://github.com/silbermm/prompt/releases/tag/v0.6.3
 [0.6.2]: https://github.com/silbermm/prompt/releases/tag/v0.6.2
 [0.6.1]: https://github.com/silbermm/prompt/releases/tag/v0.6.1
 [0.6.0]: https://github.com/silbermm/prompt/releases/tag/v0.6.0